Implement SCPort, add relevant methods to Port & SCSection

+++ modular-synth-backend/include/synth-api/section/_model/PortManager.h
@@ -26,7 +26,7 @@ namespace synth_api {
         Port* out;
 
         // TODO @bms53: Make these thread safe with locks!
-        static InputPort* getNewInputPort(Section *parent, uint64_t defaultValue);
+        static InputPort* getNewInputPort(Section *parent, float defaultValue);
         static OutputPort* getNewOutputPort(Section *parent);
 
         /*
