+++ modular-synth-backend/src/synth-api/section/PortManager.cpp
@@ -11,7 +11,7 @@
 namespace synth_api {
     std::unordered_map<Port*, Section*> PortManager::parentMap = std::unordered_map<Port*, Section*>();
 
-    InputPort *PortManager::getNewInputPort(Section *parent, uint64_t defaultValue) {
+    InputPort *PortManager::getNewInputPort(Section *parent, float defaultValue) {
         auto * inp = new InputPort(defaultValue, Rate::control);
         parentMap[inp] = parent;
         return inp;
